+++ NEWS.md
@@ -1,5 +1,8 @@
-## gDRimport 1.9.2 - 2025-11-03
+## gDRimport 1.9.3 - 2025-11-26
 * add support for Incucyte data
+
+## gDRimport 1.9.2 - 2025-11-17
+* add support for new EnVision format
 
 ## gDRimport 1.9.1 - 2025-10-30
 * synchronize Bioconductor and GitHub versioning